refactor(violations): extract useDismissedFindings to its own file
